+++ shared/src/stories/Checkbox/Checkbox.stories.js
@@ -43,13 +43,37 @@ export default {
       defaultValue: false,
       description: 'Indeterminate state for checkbox'
     },
-    valid: {
+    invalid: {
       control: 'boolean',
       defaultValue: false,
       table: {
         category: 'Props'
       },
-      description: ''
+      description: 'Invalid state'
+    },
+    label: {
+      control: 'text',
+      table: {
+        category: 'Props'
+      },
+      defaultValue: '',
+      description: 'Checkbox label'
+    },
+    helpText: {
+      control: 'text',
+      table: {
+        category: 'Props'
+      },
+      defaultValue: '',
+      description: 'Help text that can be visible below checkbox'
+    },
+    errorText: {
+      control: 'text',
+      table: {
+        category: 'Props'
+      },
+      defaultValue: '',
+      description: 'Text that displays below invalid checkbox'
     }
   }
 };
@@ -58,15 +82,26 @@ const Template = (args, { argTypes }) => ({
   components: { CheckboxBase },
   props: Object.keys(argTypes),
   template: `
-    <CheckboxBase name="checkbox" :disabled="disabled" :required="required" :indeterminate="indeterminate" :valid="valid" />`
+    <CheckboxBase name="checkbox" :label="label" :help-text="helpText" :error-text="errorText" :disabled="disabled" :required="required" :indeterminate="indeterminate" :invalid="invalid" />`
 });
 
 export const Base = Template.bind({});
+
+Base.args = {
+  invalid: false,
+  label: 'Label',
+  helpText: 'Help text',
+  errorText: 'error text'
+};
 
 const CheckboxOnRight = (args, { argTypes }) => ({
   components: { CheckboxRight },
   props: Object.keys(argTypes),
   template: `
-    <CheckboxRight name="checkbox" :disabled="disabled" :required="required" :indeterminate="indeterminate" :valid="valid" />`
+    <CheckboxRight name="checkbox" :label="label" :help-text="helpText" :error-text="errorText" :disabled="disabled" :required="required" :indeterminate="indeterminate" :invalid="invalid" />`
 });
 export const Right = CheckboxOnRight.bind({});
+
+Right.args = {
+  ...Base.args
+};