+++ boilerplates/nuxt/docs/v0.0.1/changelog.md
@@ -2,6 +2,8 @@
 
 ## 🚀 Features:
 
+- add autocomplete to Search component
+
 ## 🐛 Fixes:
 
 ## 📓 Docs: